# api-contract/openapi.yaml
openapi: 3.1.0
info:
  title: FinDoctor Client API
  version: 0.1.0
  summary: Client-facing API for financial accounts, ledger transactions, assets, liabilities, goals, tags, and analytics.
  license:
    name: Proprietary
    url: https://findoctor.example/legal/api-license
  description: |
    FinDoctor exposes user-owned financial data as explicit CRUD resources and keeps reference/admin data separate.

    Transfers are modeled as business objects connected to two underlying ledger transactions:
    one debit transaction on the source account and one credit transaction on the destination account.
    Creating, updating, or deleting a transfer must happen transactionally with those ledger rows.
servers:
  - url: http://localhost:8000
    description: Local development
  - url: https://api.findoctor.example
    description: Production placeholder
security:
  - bearerAuth: []
...
paths:
...
  /financial-institutions:
    get:
      tags: [Financial Institutions]
      summary: Search financial institutions
      operationId: listFinancialInstitutions
      parameters:
        - $ref: "#/components/parameters/SearchQuery"
        - name: country
          in: query
          schema:
            type: string
            minLength: 2
            maxLength: 2
          description: ISO 3166-1 alpha-2 country code.
        - name: providerTypeCode
          in: query
          schema:
            type: string
          description: Filter institutions by provider role code.
        - name: activeOnly
          in: query
          schema:
            type: boolean
            default: true
      responses:
        "200":
          description: Matching institutions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FinancialInstitutionList"
    post:
      tags: [Financial Institutions]
      summary: Suggest or create a financial institution
      description: Client-facing creation can be moderated server-side; admin implementations may publish directly.
      operationId: createFinancialInstitution
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FinancialInstitutionCreate"
      responses:
        "201":
          description: Created institution
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FinancialInstitution"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationError"
  /financial-institutions/{institutionId}:
    get:
      tags: [Financial Institutions]
      summary: Get financial institution
      operationId: getFinancialInstitution
      parameters:
        - $ref: "#/components/parameters/InstitutionId"
      responses:
        "200":
          description: Institution
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FinancialInstitution"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Financial Institutions]
      summary: Update financial institution metadata
      operationId: updateFinancialInstitution
      parameters:
        - $ref: "#/components/parameters/InstitutionId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FinancialInstitutionUpdate"
      responses:
        "200":
          description: Updated institution
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FinancialInstitution"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
...
  /liabilities:
    get:
      tags: [Liabilities]
      summary: List liabilities
      operationId: listLiabilities
      parameters:
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/LiabilityStatus"
        - name: liabilityTypeId
          in: query
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: User liabilities
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LiabilityList"
    post:
      tags: [Liabilities]
      summary: Create liability
      operationId: createLiability
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LiabilityCreate"
      responses:
        "201":
          description: Created liability
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Liability"
  /liabilities/{liabilityId}:
    get:
      tags: [Liabilities]
      summary: Get liability
      operationId: getLiability
      parameters:
        - $ref: "#/components/parameters/LiabilityId"
      responses:
        "200":
          description: Liability
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Liability"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Liabilities]
      summary: Update liability
      operationId: updateLiability
      parameters:
        - $ref: "#/components/parameters/LiabilityId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LiabilityUpdate"
      responses:
        "200":
          description: Updated liability
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Liability"
    delete:
      tags: [Liabilities]
      summary: Close or delete liability
      operationId: deleteLiability
      parameters:
        - $ref: "#/components/parameters/LiabilityId"
      responses:
        "204":
          description: Liability closed or deleted
        "409":
          $ref: "#/components/responses/Conflict"
  /liability-payments:
    get:
      tags: [Liability Payments]
      summary: List liability payments
      operationId: listLiabilityPayments
      parameters:
        - name: liabilityId
          in: query
          schema:
            type: string
            format: uuid
        - $ref: "#/components/parameters/DateFromQuery"
        - $ref: "#/components/parameters/DateToQuery"
      responses:
        "200":
          description: Liability payments
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LiabilityPaymentList"
    post:
      tags: [Liability Payments]
      summary: Create liability payment and linked transaction
      description: Creates the expense transaction if `transactionId` is omitted, then links it to the liability payment.
      operationId: createLiabilityPayment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LiabilityPaymentCreate"
      responses:
        "201":
          description: Created liability payment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LiabilityPayment"
  /liability-payments/{liabilityPaymentId}:
    get:
      tags: [Liability Payments]
      summary: Get liability payment
      operationId: getLiabilityPayment
      parameters:
        - $ref: "#/components/parameters/LiabilityPaymentId"
      responses:
        "200":
          description: Liability payment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LiabilityPayment"
    patch:
      tags: [Liability Payments]
      summary: Update liability payment
      operationId: updateLiabilityPayment
      parameters:
        - $ref: "#/components/parameters/LiabilityPaymentId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LiabilityPaymentUpdate"
      responses:
        "200":
          description: Updated liability payment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LiabilityPayment"
    delete:
      tags: [Liability Payments]
      summary: Delete liability payment
      operationId: deleteLiabilityPayment
      parameters:
        - $ref: "#/components/parameters/LiabilityPaymentId"
      responses:
        "204":
          description: Liability payment deleted
  /goals:
    get:
      tags: [Goals]
      summary: List financial goals
      operationId: listGoals
      responses:
        "200":
          description: User goals
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FinancialGoalList"
    post:
      tags: [Goals]
      summary: Create financial goal
      operationId: createGoal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FinancialGoalCreate"
      responses:
        "201":
          description: Created goal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FinancialGoal"
  /goals/{goalId}:
    get:
      tags: [Goals]
      summary: Get financial goal
      operationId: getGoal
      parameters:
        - $ref: "#/components/parameters/GoalId"
      responses:
        "200":
          description: Goal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FinancialGoal"
    patch:
      tags: [Goals]
      summary: Update financial goal
      operationId: updateGoal
      parameters:
        - $ref: "#/components/parameters/GoalId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FinancialGoalUpdate"
      responses:
        "200":
          description: Updated goal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FinancialGoal"
    delete:
      tags: [Goals]
      summary: Delete financial goal
      operationId: deleteGoal
      parameters:
        - $ref: "#/components/parameters/GoalId"
      responses:
        "204":
          description: Goal deleted
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    DateFromQuery:
      name: from
      in: query
      schema:
        type: string
        format: date
    DateToQuery:
      name: to
      in: query
      schema:
        type: string
        format: date
...
    GoalId:
      name: goalId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    InstitutionId:
      name: institutionId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    LiabilityId:
      name: liabilityId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    LiabilityPaymentId:
      name: liabilityPaymentId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    SearchQuery:
      name: q
      in: query
      schema:
        type: string
        minLength: 1
        maxLength: 200
...
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Conflict:
      description: Request conflicts with current resource state or uniqueness constraints
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ValidationError"
  schemas:
...
    CurrencyCode:
      type: string
      minLength: 3
      maxLength: 3
      pattern: "^[A-Z]{3}$"
      examples: [USD, EUR, RUB]
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
              additionalProperties: true
    ValidationError:
      allOf:
        - $ref: "#/components/schemas/Error"
    FinancialGoal:
      allOf:
        - $ref: "#/components/schemas/FinancialGoalCreate"
        - type: object
          required: [id, userId, currentAmount, createdAt, updatedAt]
          properties:
            id:
              type: string
              format: uuid
            userId:
              type: string
              format: uuid
            currentAmount:
              $ref: "#/components/schemas/MoneyAmount"
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time
    FinancialGoalCreate:
      type: object
      required: [name, targetAmount]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
        targetAmount:
          $ref: "#/components/schemas/MoneyAmount"
        currentAmount:
          $ref: "#/components/schemas/MoneyAmount"
        deadline:
          type: [string, "null"]
          format: date
        priority:
          type: [integer, "null"]
    FinancialGoalUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
        targetAmount:
          $ref: "#/components/schemas/MoneyAmount"
        currentAmount:
          $ref: "#/components/schemas/MoneyAmount"
        deadline:
          type: [string, "null"]
          format: date
        priority:
          type: [integer, "null"]
    FinancialGoalList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/FinancialGoal"
    FinancialInstitution:
      allOf:
        - $ref: "#/components/schemas/FinancialInstitutionCreate"
        - type: object
          required: [id, riskLevel, isActive, providerTypeIds]
          properties:
            id:
              type: string
              format: uuid
            riskLevel:
              $ref: "#/components/schemas/RiskLevel"
            isActive:
              type: boolean
            providerTypeIds:
              type: array
              items:
                type: string
                format: uuid
    FinancialInstitutionCreate:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
        country:
          type: [string, "null"]
          minLength: 2
          maxLength: 2
        websiteUrl:
          type: [string, "null"]
          format: uri
        logoUrl:
          type: [string, "null"]
          format: uri
        integrationKey:
          type: [string, "null"]
        providerTypeIds:
          type: array
          items:
            type: string
            format: uuid
          uniqueItems: true
    FinancialInstitutionUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
        country:
          type: [string, "null"]
          minLength: 2
          maxLength: 2
        websiteUrl:
          type: [string, "null"]
          format: uri
        logoUrl:
          type: [string, "null"]
          format: uri
        integrationKey:
          type: [string, "null"]
        riskLevel:
          $ref: "#/components/schemas/RiskLevel"
        isActive:
          type: boolean
        providerTypeIds:
          type: array
          items:
            type: string
            format: uuid
          uniqueItems: true
    FinancialInstitutionList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/FinancialInstitution"
    InterestType:
      type: string
      enum: [fixed, variable, promotional, none]
...
    Liability:
      allOf:
        - $ref: "#/components/schemas/LiabilityCreate"
        - type: object
          required: [id, userId, currentBalance, status, createdAt, updatedAt]
          properties:
            id:
              type: string
              format: uuid
            userId:
              type: string
              format: uuid
            currentBalance:
              $ref: "#/components/schemas/MoneyAmount"
            status:
              $ref: "#/components/schemas/LiabilityStatus"
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time
    LiabilityCreate:
      type: object
      required: [liabilityTypeId, name, currentBalance, currency]
      properties:
        liabilityTypeId:
          type: string
          format: uuid
        linkedAccountId:
          type: [string, "null"]
          format: uuid
        collateralAssetId:
          type: [string, "null"]
          format: uuid
        name:
          type: string
          minLength: 1
          maxLength: 200
        creditorInstitutionId:
          type: [string, "null"]
          format: uuid
        creditorName:
          type: [string, "null"]
        originalAmount:
          $ref: "#/components/schemas/MoneyAmountNullable"
        currentBalance:
          $ref: "#/components/schemas/MoneyAmount"
        currency:
          $ref: "#/components/schemas/CurrencyCode"
        interestRate:
          type: [number, "null"]
        interestType:
          oneOf:
            - $ref: "#/components/schemas/InterestType"
            - type: "null"
        minimumPaymentAmount:
          $ref: "#/components/schemas/MoneyAmountNullable"
        regularPaymentAmount:
          $ref: "#/components/schemas/MoneyAmountNullable"
        paymentDueDay:
          type: [integer, "null"]
          minimum: 1
          maximum: 31
        startDate:
          type: [string, "null"]
          format: date
        maturityDate:
          type: [string, "null"]
          format: date
    LiabilityUpdate:
      type: object
      properties:
        liabilityTypeId:
          type: string
          format: uuid
        linkedAccountId:
          type: [string, "null"]
          format: uuid
        collateralAssetId:
          type: [string, "null"]
          format: uuid
        name:
          type: string
          minLength: 1
          maxLength: 200
        creditorInstitutionId:
          type: [string, "null"]
          format: uuid
        creditorName:
          type: [string, "null"]
        originalAmount:
          $ref: "#/components/schemas/MoneyAmountNullable"
        currentBalance:
          $ref: "#/components/schemas/MoneyAmount"
        currency:
          $ref: "#/components/schemas/CurrencyCode"
        interestRate:
          type: [number, "null"]
        interestType:
          oneOf:
            - $ref: "#/components/schemas/InterestType"
            - type: "null"
        minimumPaymentAmount:
          $ref: "#/components/schemas/MoneyAmountNullable"
        regularPaymentAmount:
          $ref: "#/components/schemas/MoneyAmountNullable"
        paymentDueDay:
          type: [integer, "null"]
          minimum: 1
          maximum: 31
        startDate:
          type: [string, "null"]
          format: date
        maturityDate:
          type: [string, "null"]
          format: date
        status:
          $ref: "#/components/schemas/LiabilityStatus"
    LiabilityList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Liability"
    LiabilityStatus:
      type: string
      enum: [active, paid_off, defaulted, refinanced, closed]
...
    LiabilityPayment:
      allOf:
        - $ref: "#/components/schemas/LiabilityPaymentCreate"
        - type: object
          required: [id, userId, transactionId, createdAt]
          properties:
            id:
              type: string
              format: uuid
            userId:
              type: string
              format: uuid
            transactionId:
              type: string
              format: uuid
            createdAt:
              type: string
              format: date-time
    LiabilityPaymentCreate:
      type: object
      required: [liabilityId, accountId, paymentDate, totalAmount, currency]
      properties:
        liabilityId:
          type: string
          format: uuid
        accountId:
          type: string
          format: uuid
          description: Account used for the underlying expense transaction when transactionId is omitted.
        transactionId:
          type: [string, "null"]
          format: uuid
          description: Existing transaction to attach to this liability payment.
        recurringTransactionId:
          type: [string, "null"]
          format: uuid
        paymentDate:
          type: string
          format: date
        totalAmount:
          $ref: "#/components/schemas/MoneyAmount"
        principalAmount:
          $ref: "#/components/schemas/MoneyAmountNullable"
        interestAmount:
          $ref: "#/components/schemas/MoneyAmountNullable"
        feeAmount:
          $ref: "#/components/schemas/MoneyAmountNullable"
        currency:
          $ref: "#/components/schemas/CurrencyCode"
        balanceAfterPayment:
          $ref: "#/components/schemas/MoneyAmountNullable"
    LiabilityPaymentUpdate:
      type: object
      properties:
        recurringTransactionId:
          type: [string, "null"]
          format: uuid
        paymentDate:
          type: string
          format: date
        totalAmount:
          $ref: "#/components/schemas/MoneyAmount"
        principalAmount:
          $ref: "#/components/schemas/MoneyAmountNullable"
        interestAmount:
          $ref: "#/components/schemas/MoneyAmountNullable"
        feeAmount:
          $ref: "#/components/schemas/MoneyAmountNullable"
        currency:
          $ref: "#/components/schemas/CurrencyCode"
        balanceAfterPayment:
          $ref: "#/components/schemas/MoneyAmountNullable"
    LiabilityPaymentList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/LiabilityPayment"
...
    MoneyAmount:
      type: string
      pattern: "^-?\\d+(\\.\\d{1,8})?$"
      description: Decimal amount encoded as a string to avoid floating point loss.
      examples: ["1250.00"]
    MoneyAmountNullable:
      type: [string, "null"]
      pattern: "^-?\\d+(\\.\\d{1,8})?$"
...
    RiskLevel:
      type: string
      enum: [low, medium, high, unknown]
...